test(settings): move validate-only tests to correctly named files
